+++ lua/fzfx.lua
@@ -32,8 +32,8 @@ local function setup(opts)
   require("fzfx.detail.module").setup()
   require("fzfx.detail.rpcserver").setup()
   require("fzfx.detail.yanks").setup()
-  require("fzfx.detail.bat_helpers").setup()
   require("fzfx.detail.popup").setup()
+  require("fzfx.detail.bat_helpers").setup()
 
   local general = require("fzfx.detail.general")
